land b1_4_smallfix on b1_4(20050202_1817)
- readahead could read extra page into cache that wasn't ejected (5388)
- prevent races in class_attach/setup/cleanup/detach (5260)
- don't dereference de->d_inode after l_dput of de (5458)
- use "int" for stripe value returned from lock_to_stripe (5544)
- mballoc allocation and error-checking fixes in 2.6 (5504)
- block device patches to fix I/O request sizes in 2.6 (5482)
- protect KMS changes with its own lock, not i_sem (5492, 5624, 3453)
- look up hostnames for IB nals (5602)
- fix mount usage message, return errors per mount(8) (5168)
- remove ancient LOV_MAGIC_V0 EA support (5047)
- add "disk I/Os in flight" and "I/O req time" stats in obdfilter

+++ lustre/include/linux/lustre_net.h
@@ -101,7 +101,7 @@
  * considered full when less than ?_MAXREQSIZE is left in them.
  */
 
-#define LDLM_NUM_THREADS        min(smp_num_cpus * smp_num_cpus * 8, 64)
+#define LDLM_NUM_THREADS        min((int)(smp_num_cpus * smp_num_cpus * 8), 64)
 #define LDLM_NBUFS       64
 #define LDLM_BUFSIZE    (8 * 1024)
 #define LDLM_MAXREQSIZE (5 * 1024)
@@ -221,7 +221,7 @@ struct ptlrpc_request_set {
         wait_queue_head_t *set_wakeup_ptr;
         struct list_head  set_requests;
         set_interpreter_func    set_interpret; /* completion callback */
-        void             *set_arg; /* completion context */
+        void              *set_arg; /* completion context */
         /* locked so that any old caller can communicate requests to
          * the set holder who can then fold them into the lock-free set */
         spinlock_t        set_new_req_lock;
